requirements.txt
pyyaml>=6.0
python-dotenv>=1.0.0
